Code-dump of last known version of WonderBrush and its dependencies
There is no version history preserved in this repository. It could be
possible to retrieve it from the YellowBites SVN server, which may still be
running. However, it contains personal information from WonderBrush users
who have purchased a license, as such, I'll probably make no efforts to
fix the history.
I have not yet tried to build this code, but it is the version which was
used to build the WonderBrush that currently ships with Haiku.