Work around runtime_loader bug.
This works on GCC2, but on GCC7 it seems the agg symbols from WonderBrush
get used instead of the agg symbols inside libbe (libicon), which they
really shouldn't because the libicon ones should be hidden and non-weak.
So I don't know what's going on there, but this works around it for now.