Fix crash using pen tool on 64-bit version
When drawing outside the canvas, the offset
into the bitmap buffer is calculated using a
negative value for Y, the calculation was
wrong when "bytes per row" was an unsigned type,
since it buffer address needs to point before
the buffer in that case. There will be no drawing
at that "outside" location, but at the edge
case of the "pen area" overlapping with the bitmap
top, the remaining part (left-bottom and right-bottom)
are still drawn after shifting the bitmap offset
into the valid region with "+= bpr - 1". It's at
that point were it crashed. Not sure if this
also affected the 32-bit version, only without
showing right away. There is a good chance it
corrupted memory.