Fix build with modern gcc
The -I- option seems broken in current versions of gcc, and prevents
searching for includes located in the same directory as the file being
processed (which should be the first thing searched).

Note that the option has been deprecated for years, and is kept only
because of gcc2 support.

Switch Wonderbrush's jamfiles to use the new -iquote system instead,
when compiled on 64-bit Haiku. Unfortunately, just changing the
variables defined in the existing Jamfile was not enough, because
FInclude is implemented in Jambase and hardcodes -I. So, an extra
override for FIncludes is also added.

I only tested this change on 64bit Haiku.